+++ asyncz/schedulers/types.py
@@ -1,15 +1,16 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any, Callable, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Tuple, Union
 
 from asyncz.events.constants import (
     ALL_EVENTS,
 )
-from asyncz.typing import UndefinedType, undefined
+from asyncz.typing import Undefined, undefined
 
 if TYPE_CHECKING:
     from datetime import datetime
+    from logging import Logger
     from threading import RLock
 
     from asyncz.events.base import SchedulerEvent
@@ -19,7 +20,22 @@ if TYPE_CHECKING:
     from asyncz.triggers.types import TriggerType
 
 
+class LoggersType(ABC):
+    def __init__(self) -> None:
+        self.data: Dict[str, Logger] = {}
+
+    @abstractmethod
+    def __missing__(self, item: str) -> Logger: ...
+
+    def __getitem__(self, item: str) -> Logger:
+        if item not in self.data:
+            self.data[item] = self.__missing__(item)
+        return self.data[item]
+
+
 class SchedulerType(ABC):
+    loggers: LoggersType
+
     @abstractmethod
     def start(self, paused: bool = False) -> bool:
         """
@@ -130,12 +146,12 @@ class SchedulerType(ABC):
         kwargs: Optional[Any] = None,
         id: Optional[str] = None,
         name: Optional[str] = None,
-        mistrigger_grace_time: Union[int, UndefinedType, None] = undefined,
-        coalesce: Union[bool, UndefinedType] = undefined,
-        max_instances: Union[int, UndefinedType, None] = undefined,
-        next_run_time: Union[datetime, str, UndefinedType, None] = undefined,
-        store: str = "default",
-        executor: str = "default",
+        mistrigger_grace_time: Union[int, Undefined, None] = undefined,
+        coalesce: Union[bool, Undefined] = undefined,
+        max_instances: Union[int, Undefined, None] = undefined,
+        next_run_time: Union[datetime, str, Undefined, None] = undefined,
+        store: Optional[str] = None,
+        executor: Optional[str] = None,
         replace_existing: bool = False,
         **trigger_args: Any,
     ) -> TaskType: