+++ asyncz/stores/sqlalchemy.py
@@ -95,7 +95,9 @@ class SQLAlchemyStore(BaseStore):
                     tasks.append(self.rebuild_task(row.state))
                 except Exception:
                     task_id = row.id
-                    self.logger.exception(f"Unable to restore task '{task_id}'. Removing it...")
+                    cast("SchedulerType", self.scheduler).loggers[self.logger_name].exception(
+                        f"Unable to restore task '{task_id}'. Removing it..."
+                    )
                     failed_task_ids.append(task_id)
 
             if failed_task_ids: