+++ docs_src/tasks/create_task.py
@@ -22,3 +22,16 @@ Task(
 )
 
 scheduler.start()
+
+# or manually submit one with a changed trigger
+
+t = Task(
+    id="my-task2",
+    fn=check_status,
+    max_instances=1,
+    coalesce=True,
+)
+
+scheduler.add_task(
+    t, trigger=CronTrigger(day_of_week="mon,tue,wed,thu,fri,sat,sun", hour=8, minute=1)
+)